criando o produto service

+++ src/main/java/com/deloitte/model/Pedido.java
@@ -33,7 +33,7 @@ public class Pedido {
 
     private Integer quantidade;
 
-    private BigDecimal valorTotal;
+    private Double valorTotal;
 
 
 }
